Unify all `Result`s and `ErrorCode`s
