+++ Sts1CobcSw/CobcSoftware/EduListenerThread.cpp
@@ -7,6 +7,7 @@
 #include <Sts1CobcSw/Edu/ProgramStatusHistory.hpp>
 #include <Sts1CobcSw/Edu/Types.hpp>
 #include <Sts1CobcSw/Hal/IoNames.hpp>
+#include <Sts1CobcSw/Outcome/Outcome.hpp>
 #include <Sts1CobcSw/Utility/RodosTime.hpp>
 
 #include <strong_type/difference.hpp>
