+++ Sts1CobcSw/FileSystem/LfsWrapper.hpp
@@ -1,8 +1,8 @@
 #pragma once
 
 
-#include <Sts1CobcSw/FileSystem/ErrorsAndResult.hpp>
 #include <Sts1CobcSw/FileSystem/LfsMemoryDevice.hpp>
+#include <Sts1CobcSw/Outcome/Outcome.hpp>
 #include <Sts1CobcSw/Serial/Byte.hpp>
 
 #include <littlefs/lfs.h>
