+++ Tests/ManualTests/CMakeLists.txt
@@ -20,6 +20,7 @@ if(CMAKE_SYSTEM_NAME STREQUAL Generic)
                 strong_type::strong_type
                 Sts1CobcSw_Edu
                 Sts1CobcSw_Hal
+                Sts1CobcSw_Outcome
                 Sts1CobcSw_ProgramId
                 Sts1CobcSw_Serial
                 Sts1CobcSw_Utility
