+++ Tests/UnitTests/LfsWrapper.test.cpp
@@ -1,11 +1,11 @@
 #include <Tests/CatchRodos/TestMacros.hpp>
 
-#include <Sts1CobcSw/FileSystem/ErrorsAndResult.hpp>
 #include <Sts1CobcSw/FileSystem/LfsMemoryDevice.hpp>
 #ifdef __linux__
     #include <Sts1CobcSw/FileSystem/LfsRam.hpp>
 #endif
 #include <Sts1CobcSw/FileSystem/LfsWrapper.hpp>
+#include <Sts1CobcSw/Outcome/Outcome.hpp>
 #include <Sts1CobcSw/Serial/Byte.hpp>
 #include <Sts1CobcSw/Utility/Span.hpp>
